Moved controller function for title out to a new controller
