+++ src/main/twirl/views/manage/opportunityOverview.scala.html
@@ -63,7 +63,7 @@ backLink=Some(BackLink("Opportunity library", controllers.manage.routes.Opportun
                                 <tr>
                                     <td>
                                         1. <a id="op-setup-link1"
-                                              href="@controllers.manage.routes.OpportunityController.viewTitle(opportunity.id)"
+                                              href="@controllers.manage.routes.TitleController.view(opportunity.id)"
                                             > Opportunity title</a>
                                     </td>
                                 </tr>
